♻️ refactor: stockRepository 파일 위치를 /src/stock 에서 /src/stock/repository 로 변경
- 바뀐 위치에 따라 import 업데이트
